requirements-gpu.txt
# ========================================
# AI-EVA Demo GPU 版本依赖配置
# ========================================
# 这是 GPU 加速版本的依赖配置
# 需要 NVIDIA GPU 和 CUDA 支持
# ========================================

# ========== 重要说明 ==========
# 1. 先安装 PyTorch GPU 版本：
#    pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
#
# 2. 然后安装此文件的其他依赖：
#    pip install -r requirements-gpu.txt
#
# 3. 根据你的 CUDA 版本选择：
#    - CUDA 11.8: cu118
#    - CUDA 12.1: cu121
#    访问 https://pytorch.org/ 获取最新信息
# ========================================

# 核心框架
fastapi>=0.111.1
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# AI 模型核心依赖
numpy>=1.24.0,<2.0.0
scipy>=1.10.0

# ChatTTS 语音合成
ChatTTS
omegaconf>=2.3.0
vocos
vector-quantize-pytorch

# SenseVoice 语音识别
funasr>=1.1.3
modelscope>=1.9.0
huggingface_hub>=0.19.0

# 音频处理
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.1
audioread>=3.0.0

# 工具库
python-dotenv>=1.0.0
aiofiles>=23.0.0
pyyaml>=6.0

# HTTP 客户端
httpx>=0.25.0
requests>=2.31.0
aiohttp>=3.9.0

# 性能与监控
psutil>=5.9.0
cachetools>=5.3.0

# 日志与调试
loguru>=0.7.0
colorama>=0.4.6

# 数据处理
pandas>=2.0.0
pillow>=10.0.0

# 其他依赖
tqdm>=4.65.0
typing-extensions>=4.8.0
pydantic-settings>=2.0.0

# ========================================
# GPU 加速安装完整步骤
# ========================================
# 1. 卸载现有的 PyTorch (如果有)：
#    pip uninstall torch torchaudio torchvision
#
# 2. 安装 PyTorch GPU 版本 (CUDA 11.8)：
#    pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
#
# 3. 验证 CUDA 是否可用：
#    python -c "import torch; print(torch.cuda.is_available())"
#
# 4. 安装其他依赖：
#    pip install -r requirements-gpu.txt
#
# 5. 设置环境变量：
#    在 .env 文件中设置 USE_GPU=true
# ========================================
